backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Redis
redis==5.2.1

# Auth
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Validation
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Celery
celery[redis]==5.4.0

# Socket.IO
python-socketio==5.12.1

# Kite Connect
kiteconnect==5.0.1

# Data
pandas==2.2.3
numpy==2.2.1

# Encryption
cryptography==44.0.0

# Security
RestrictedPython==7.4

# Rate limiting
slowapi==0.1.9

# HTTP client
httpx==0.28.1

# Utilities
python-dateutil==2.9.0
pytz==2024.2
